PMM-7513 change severity when switching template
